Refactor layout and authentication handling; update navbar and home page UI
- Changed import path for NavBar component to match casing.
- Improved authentication status fetching by using Headers for cookie management.
- Added error handling for authentication status response.
- Updated footer layout in RootLayout for better structure.
- Simplified Home component by removing unused token exchange logic.
- Enhanced button layout in Home component for better accessibility.
- Removed unused Period page component.
- Cleaned up Roulette page by removing unnecessary imports.
- Added Tracker page component for period tracking.
- Introduced AddDDayDialog component for creating D-Days.
- Created CalendarGrid component for displaying calendar days.
- Implemented CalendarHeader component for month navigation.
- Developed DDayIndicator component for displaying D-Day events.
- Created DDaySheet component for listing D-Days in a sheet format.
- Added useCalendar and useDDays hooks for calendar and D-Day logic.
