+++ frontend/app/layout.tsx
@@ -6,7 +6,7 @@ import { cookies } from "next/headers";
 
 import { ThemeProvider } from "@/components/theme-provider";
 import { AuthProvider } from "@/components/auth-provider";
-import { NavBar } from "@/components/navbar";
+import { NavBar } from "@/components/Navbar";
 
 const BACKEND_URL = process.env.NEXT_PUBLIC_BACKEND_URL;
 
@@ -24,14 +24,29 @@ export const metadata: Metadata = {
 async function getAuthStatus() {
     const cookieStore = await cookies();
     const sessionCookie = cookieStore.get("calple_session");
+
     if (!sessionCookie) return { isAuthenticated: false, user: null };
 
     try {
+        const headers = new Headers();
+        headers.append(
+            "Cookie",
+            `${sessionCookie.name}=${sessionCookie.value}`
+        );
+        headers.append("Content-Type", "application/json");
+
         const response = await fetch(`${BACKEND_URL}/api/auth/status`, {
-            headers: { Cookie: `calple_session=${sessionCookie.value}` },
+            headers,
             cache: "no-store",
         });
+
+        if (!response.ok) {
+            console.log("Auth status error:", response.status);
+            return { isAuthenticated: false, user: null };
+        }
+
         const data = await response.json();
+        console.log("Auth status response:", data);
         return {
             isAuthenticated: data.authenticated,
             user: data.user,
@@ -60,22 +75,24 @@ export default async function RootLayout({
                         <NavBar />
                         {children}
                         <div className="fixed bottom-0 left-0 w-full flex justify-center z-50 border-t border-dashed px-8">
-                            <div className="flex flex-col px-6 py-2 container text-muted-foreground text-xs">
-                                J&E &copy; 2025
-                            </div>
-                            <div className="flex justify-end gap-2 px-6 py-2 container text-muted-foreground text-xs">
-                                <a
-                                    href="/privacy"
-                                    className="text-muted-foreground hover:text-stone-500 underline"
-                                >
-                                    [Privacy Policy]
-                                </a>
-                                <a
-                                    href="/terms"
-                                    className="text-muted-foreground hover:text-stone-500 underline"
-                                >
-                                    [Terms of Service]
-                                </a>
+                            <div className="container flex">
+                                <div className="flex flex-col px-6 py-2 container text-muted-foreground text-xs">
+                                    J&E &copy; 2025
+                                </div>
+                                <div className="flex justify-end gap-2 px-6 py-2 container text-muted-foreground text-xs">
+                                    <a
+                                        href="/privacy"
+                                        className="text-muted-foreground hover:text-stone-500 underline"
+                                    >
+                                        [Privacy]
+                                    </a>
+                                    <a
+                                        href="/terms"
+                                        className="text-muted-foreground hover:text-stone-500 underline"
+                                    >
+                                        [Terms]
+                                    </a>
+                                </div>
                             </div>
                         </div>
                     </AuthProvider>
