--- frontend/app/period/page.tsx
@@ -1,7 +1,0 @@
-import { useState, useEffect } from "react";
-import { login, logout } from "@/lib/utils";
-import { useTheme } from "next-themes";
-
-export default function Period() {
-    return <div className="p-8 text-2xl">Period Page</div>;
-}
