+++ frontend/app/roulette/page.tsx
@@ -1,4 +1,3 @@
-import { useState, useEffect } from "react";
 import { login, logout } from "@/lib/utils";
 import { useTheme } from "next-themes";
