+++ frontend/app/tracker/page.tsx
@@ -1,0 +1,6 @@
+import { login, logout } from "@/lib/utils";
+import { useTheme } from "next-themes";
+
+export default function Tracker() {
+    return <div className="p-8 text-2xl">Period Tracker Page</div>;
+}
