+++ frontend/hooks/useCalendar.tsx
@@ -1,0 +1,106 @@
+import { useState } from "react";
+
+export function useCalendar() {
+    const [currentDate, setCurrentDate] = useState(new Date());
+    const [selectedDate, setSelectedDate] = useState(new Date());
+
+    const getDaysInMonth = (year: number, month: number) => {
+        return new Date(year, month + 1, 0).getDate();
+    };
+
+    const getFirstDayOfMonth = (year: number, month: number) => {
+        return new Date(year, month, 1).getDay();
+    };
+
+    const generateMonthData = () => {
+        const year = currentDate.getFullYear();
+        const month = currentDate.getMonth();
+
+        const daysInMonth = getDaysInMonth(year, month);
+        const firstDayOfMonth = getFirstDayOfMonth(year, month);
+
+        const days = [];
+        for (let i = 0; i < firstDayOfMonth; i++) {
+            days.push(null);
+        }
+
+        for (let i = 1; i <= daysInMonth; i++) {
+            days.push(i);
+        }
+
+        return days;
+    };
+
+    const calculateRequiredRows = () => {
+        const year = currentDate.getFullYear();
+        const month = currentDate.getMonth();
+        const firstDayOfMonth = getFirstDayOfMonth(year, month);
+        const daysInMonth = getDaysInMonth(year, month);
+        const totalCells = firstDayOfMonth + daysInMonth;
+        return Math.ceil(totalCells / 7);
+    };
+
+    const formatMonth = (date: Date) => {
+        return date.toLocaleString("default", {
+            month: "short",
+            year: "numeric",
+        });
+    };
+
+    const goToNextMonth = () => {
+        const newDate = new Date(currentDate);
+        newDate.setMonth(newDate.getMonth() + 1);
+        setCurrentDate(newDate);
+    };
+
+    const goToPrevMonth = () => {
+        const newDate = new Date(currentDate);
+        newDate.setMonth(newDate.getMonth() - 1);
+        setCurrentDate(newDate);
+    };
+
+    const goToToday = () => {
+        const today = new Date();
+        setCurrentDate(today);
+        setSelectedDate(today);
+    };
+
+    const isToday = (day: number | null) => {
+        if (!day) return false;
+        const today = new Date();
+        return (
+            day === today.getDate() &&
+            currentDate.getMonth() === today.getMonth() &&
+            currentDate.getFullYear() === today.getFullYear()
+        );
+    };
+
+    const isSelected = (day: number | null) => {
+        if (!day) return false;
+        return (
+            day === selectedDate.getDate() &&
+            currentDate.getMonth() === selectedDate.getMonth() &&
+            currentDate.getFullYear() === selectedDate.getFullYear()
+        );
+    };
+
+    const selectDate = (day: number) => {
+        const newDate = new Date(currentDate);
+        newDate.setDate(day);
+        setSelectedDate(newDate);
+    };
+
+    return {
+        currentDate,
+        selectedDate,
+        monthData: generateMonthData(),
+        requiredRows: calculateRequiredRows(),
+        formatMonth,
+        goToNextMonth,
+        goToPrevMonth,
+        goToToday,
+        isToday,
+        isSelected,
+        selectDate,
+    };
+}
